Broke chatbot down into modules
